+++ src/main/java/uk/ac/ebi/embl/fastareader/sequenceutils/SequenceIndex.java
@@ -10,7 +10,6 @@
  */
 package uk.ac.ebi.embl.fastareader.sequenceutils;
 
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -23,7 +22,7 @@ public final class SequenceIndex {
     public long endNBasesCount;
     private final List<LineEntry> lines;
     public long nextHeaderByte; // byte offset of next '>' at line start, or fileSize (EOF)
-    public BigInteger totalNBases;
+    public long totalNBases;
 
     public SequenceIndex(
             long firstBaseByte,
@@ -32,7 +31,7 @@ public final class SequenceIndex {
             long endNBasesCount,
             List<LineEntry> lines,
             long nextHeader,
-            BigInteger totalNBases) {
+            long totalNBases) {
         this.firstBaseByte = firstBaseByte;
         this.startNBasesCount = startNBasesCount;
         this.lastBaseByte = lastBaseByte;
