+++ src/test/java/uk/ac/ebi/embl/fastareader/sequenceutils/SequenceIndexBuilderTest.java
@@ -13,7 +13,6 @@ package uk.ac.ebi.embl.fastareader.sequenceutils;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.io.IOException;
-import java.math.BigInteger;
 import java.nio.channels.FileChannel;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
@@ -110,7 +109,7 @@ public class SequenceIndexBuilderTest {
             assertEquals(2, idx.startNBasesCount, "leading Ns only from first sequence line");
             assertEquals(2, idx.endNBasesCount, "trailing Ns only from last sequence line");
             // assert total count of N bases is correctly done
-            assertEquals(BigInteger.valueOf(4), idx.totalNBases);
+            assertEquals(4, idx.totalNBases);
 
             // nextHeaderByte should point to '>' of NEXT header
             long expectedNextHeader = header.length() + l1.length() + l2.length() + l3.length() + empties.length();
@@ -143,7 +142,7 @@ public class SequenceIndexBuilderTest {
             // no trailing Ns on last line (all 'g')
             assertEquals(0, idx.endNBasesCount);
             // assert total count of N bases is correctly done
-            assertEquals(BigInteger.valueOf(2), idx.totalNBases);
+            assertEquals(2, idx.totalNBases);
 
             // nextHeader should be at the '>' byte of H2
             long expectedNext = fasta.lastIndexOf(">H2\n"); // ascii index
@@ -176,7 +175,7 @@ public class SequenceIndexBuilderTest {
             // trailing Ns until the start
             assertEquals(33, idx.endNBasesCount);
             // assert total count of N bases is correctly done
-            assertEquals(BigInteger.valueOf(idx.totalBases()), idx.totalNBases);
+            assertEquals(idx.totalBases(), idx.totalNBases);
 
             // nextHeader should be at the '>' byte of H2
             long expectedNext = fasta.lastIndexOf(">H2\n"); // ascii index
@@ -213,7 +212,7 @@ public class SequenceIndexBuilderTest {
             assertEquals(1, idx.startNBasesCount, "only first line leading Ns");
             assertEquals(1, idx.endNBasesCount, "only last line trailing Ns");
             // assert total count of N bases is correctly done
-            assertEquals(BigInteger.valueOf(6), idx.totalNBases);
+            assertEquals(6, idx.totalNBases);
 
             // Middle line of Ns shouldn't change edge counts
             assertEquals(idx.linesView().get(1).lengthBases(), 4);
